Remove support for old Go versions
Fixes #51
