+++ go.mod
@@ -1,1 +1,3 @@
+go 1.16
+
 module github.com/fsnotify/fsevents
