--- go_1_9_2_after.go
@@ -1,12 +1,0 @@
-//go:build darwin && go1.9.2
-// +build darwin,go1.9.2
-
-package fsevents
-
-/*
-#include <CoreServices/CoreServices.h>
-*/
-import "C"
-
-// eventIDSinceNow is a sentinel to begin watching events "since now".
-const eventIDSinceNow = uint64(C.kFSEventStreamEventIdSinceNow)
